requirements.txt
flask
flask-cors
flask-socketio
opencv-python
numpy
pillow
python-socketio
